+++ samples/corda/corda-simple-application/clients/src/main/kotlin/com/cordaSimpleApplication/client/ConfigureNetwork.kt
@@ -25,6 +25,8 @@ import net.corda.core.messaging.startFlow
 import org.json.JSONObject
 import java.util.*
 
+import com.weaver.corda.sdk.CredentialsExtractor
+
 /**
  * TODO: Documentation
  */
